domain_parameterへの詰め替え処理について、mac版でエラーが出る不具合 & ompのリンクが設定されていない不具合を修正

+++ CS_server/VisModule/Visualization/Object/CS_PointObjectGenerator.cpp
@@ -430,7 +430,7 @@ vismodule::PointObject* CS_PointObjectGenerator::sampling( const Argument& param
             raw_values[j] = values[j].get();
         }
 
-        int resol[3] = { svo_p->resolution().x(), svo_p->resolution().y(), svo_p->resolution().z()};
+        int resol[3] = { static_cast<int>(svo_p->resolution().x()), static_cast<int>(svo_p->resolution().y()), static_cast<int>(svo_p->resolution().z())};
         domain_parameters_struct dom={
          volume.minObjectCoord().x()
         ,volume.minObjectCoord().y()
