+++ CS_server/VisModule/Visualization/Object/GlyphSeedGenerator.cpp
@@ -340,7 +340,7 @@ void GlyphSeedGenerator::sampling( vismodule::VolumeObjectBase* volume,const jpv
     {
         const vismodule::StructuredVolumeObject* svo_p = static_cast<const vismodule::StructuredVolumeObject*>( volume );
 
-        int resol[3] = { svo_p->resolution().x(), svo_p->resolution().y(), svo_p->resolution().z()};
+        int resol[3] = { static_cast<int>(svo_p->resolution().x()), static_cast<int>(svo_p->resolution().y()), static_cast<int>(svo_p->resolution().z())};
         domain_parameters_struct dom={
          volume->minObjectCoord().x()
         ,volume->minObjectCoord().y()
